Added slice_equal and convert_to_array
The existing array_equal code did not take into account containers of
different sizes. The equivalence checks overloads funnel to this function
as a precaution, so that zip().all() isn't being done repeatedly.

The convert_to_array macro replaces the existing versions, so that an error
is returned if the container is not a matching size. The error checking was
added to every type when being deserialised.

NOTE: slice_equal could be written more abstractly to "range_equal", and
accept iterables directly.